chore(🦉): [Add .gitignore and .npmignore files]
## Chores
- Create a .gitignore file to exclude build, distribution, and environment files.
- Create a .npmignore file to specify files and directories to ignore in npm packages.

## Description
This commit introduces two new files: .gitignore and .npmignore. The .gitignore file ensures that unnecessary files such as build artifacts and environment configurations are not tracked by Git. The .npmignore file specifies which files should be excluded when publishing the package to npm, helping to keep the package clean and lightweight.